Re-organize and clean up some Unix Drawing code.
